Move fetcher client into catchup (#774)
* changes.

* adding dialer.

* Move fetcher client into catchup, step 1.
( most unit tests are still broken )

* DRAFT: using channel to offload the mutex.

* Taking care of the lock triggering deadlock detection.

* cleaning unnecessary changes.

* cleaning unnecessary changes.

* minor fixes

* GetNetTransport modifying returning copy of the http.Transport

* workaround to avoid the race detection trigger.

* Testing a different approach to ovrride the Dial/DialContext by embedding the default transport into another object instead of changing the default transport.

* Adding RateLimitedTransport to wrap around the http.Transport

* fixing lint

* Separating Dialer from Transport, initializing the Dialer and Transport params (timeout, etc)

* changes.

* update.

* fix few more unit tests.

* fix syncer tests.

* undo change.

* Add a comment.
