+++ rpcs/httpTxSync.go
@@ -48,7 +48,9 @@ type HTTPTxSync struct {
 
 const requestContentType = "application/x-www-form-urlencoded"
 
-func responseBytes(response *http.Response, log logging.Logger, limit uint64) (data []byte, err error) {
+// ResponseBytes reads the content of the response object and return the body content
+// while obeying the read size limits
+func ResponseBytes(response *http.Response, log logging.Logger, limit uint64) (data []byte, err error) {
 	// response.Body is always non-nil
 	defer response.Body.Close()
 	if response.ContentLength >= 0 {
@@ -159,7 +161,7 @@ func (hts *HTTPTxSync) Sync(ctx context.Context, bloom *bloom.Filter) (txgroups 
 		return nil, fmt.Errorf("txSync POST invalid content type '%s'", contentTypes[0])
 	}
 
-	data, err := responseBytes(response, hts.log, hts.maxTxSyncResponseBytes)
+	data, err := ResponseBytes(response, hts.log, hts.maxTxSyncResponseBytes)
 	if err != nil {
 		hts.log.Warn("txSync body read failed: ", err)
 		return nil, err
